move edge filtering to events
